I've pretty much changed everything
generally, all the models are now in gdcf_model and all the other crates depend on that. that leads to the actual gdcf crate to only define traits for cache/api client and the code to connect them together, nothing else (well it also contains the request structs, but they kinda belong there). GDCF parse only depends on gdcf_model.

The idea behind this is that you can now easily depend on gdcf_model/gdcf_parse without dragging in a shit load of dependencies from tokio and friends by depending on gdcf (which was required previously)