SafeUnparse support
SafeUnparse is a wrapper struct that keeps track of all index/value pairs that got ignored during parsing and ensures they will still be contained when unparsing (something Parse impls in general do not need to guarantee). We now have a lifetime on the Parse struct, but that's actually better because now the parse result can use references into the data source for even more efficient parsing (if it wants. gdcf doesn't make use of this outside of SafeUnparse yet)