Rewrite of the entire core mechanics
Mainly this mean we no longer differentiate between vectors of objects and individual objects at the future level, as this is now all handled via clever `Upgradable` implementations. The Upgradable interface has also been greatly simplified and no longer causes a ton of superfluous requests (It used to ignore the fact that some objects can be returned by different requests and instead demanded that a specific request result is cached instead of just a specific object). Additionally, the mess that was GdcfFuture has been broken down in two smaller traits that do not require reimplementation of a poll methods (this is mainly done in preparation for the switch to futures 0.3). The only reason for that trait to exist was to make GdcfStream work. It was very badly designed (the usage it was intended for caused cycles during compilation). Along with this, GdcfStream has obviously been largely rewritten